requirements.txt
# Discord
discord.py>=2.3.0

# Database
asyncpg>=0.29.0

# Web Scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3

# Logging (built-in, but specifying for clarity)
# logging - standard library
